fix(materials): show resume rationale in apply review
